+++ R/gizmo-density.R
@@ -195,9 +195,9 @@ GizmoDensity <- ggproto(
   setup_elements = function(params, elements, theme) {
     theme$legend.frame <- theme$legend.frame %||% element_blank()
     if (params$direction == "horizontal") {
-      theme$legend.key.width  <- rel(5)
+      theme$legend.key.width  <- (theme$legend.key.width %||% rel(1)) * 5
     } else {
-      theme$legend.key.height <- rel(5)
+      theme$legend.key.height <- (theme$legend.key.height %||% rel(1)) * 5
     }
     Guide$setup_elements(params, elements, theme)
   },