+++ NEWS.md
@@ -1,5 +1,7 @@
 # legendry (development version)
 
+* `legend.key.width/height` is no longer ignored in `guide_colbar()` (#81).
+
 # legendry 0.2.2
 
 This is a patch release without new features, improvements or bug fixes.